+++ src/test/php/net/daringfireball/markdown/unittest/ParsingTest.class.php
@@ -6,6 +6,9 @@ use net\daringfireball\markdown\Text;
 use net\daringfireball\markdown\Table;
 use net\daringfireball\markdown\Row;
 use net\daringfireball\markdown\Cell;
+use net\daringfireball\markdown\Image;
+use net\daringfireball\markdown\Link;
+use net\daringfireball\markdown\NodeList;
 
 class ParsingTest extends MarkdownTest {
 
@@ -26,6 +29,26 @@ class ParsingTest extends MarkdownTest {
         new Paragraph([new Text('Hello World')])
       ]),
       $this->fixture->parse('Hello World')
+    );
+  }
+
+  #[@test]
+  public function link() {
+    $this->assertEquals(
+      new ParseTree([
+        new Paragraph([new Link('url', new NodeList([new Text('link')]))])
+      ]),
+      $this->fixture->parse('[link](url)')
+    );
+  }
+
+  #[@test]
+  public function image() {
+    $this->assertEquals(
+      new ParseTree([
+        new Paragraph([new Image('url', new NodeList([new Text('image')]))])
+      ]),
+      $this->fixture->parse('![image](url)')
     );
   }